requirements.txt
arxiv
requests
python-dotenv
psutil
